Update makeBoxplot.R
I extracted the data and samples from the preprocessBEclear function, and the samples table is a data.table data.frame. There are too many samples in my data to view by sample, so I opted for bySamples = F. I received the error: "Error in data[, s] : subscript out of bounds" when I tried to run makeBoxplot(data, samples, score.table, bySamples = F). Debugging from the source code, the problem appears to be at line 112, where "sample_id" is in quotations. As is, a list is returned that looked like this: c(\"sample_1\", \"sample_2\",..., \"sample_n\"), and when trying to subset the data matrix at line 114 with the  boxplotData for-loop the error: "Error in data[, Batch_1] : invalid subscript type 'list'" occurs. Removing the quotations from "sample_id" in line 112 returns a vector of characters that looks like "sample_1", "sample_2", ... , "sample_n". This vector correctly subset the data matrix.